Moved calling code interface to the field class, improved demo app, added flag to force local phone number
