+++ gemsfx-demo/src/main/java/com/dlsc/gemsfx/demo/PhoneNumberFieldApp.java
@@ -1,18 +1,21 @@
 package com.dlsc.gemsfx.demo;
 
-import com.dlsc.gemsfx.CountryCallingCode;
 import com.dlsc.gemsfx.PhoneNumberField;
 import javafx.application.Application;
+import javafx.beans.InvalidationListener;
 import javafx.beans.binding.Bindings;
 import javafx.beans.value.ObservableValue;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
 import javafx.scene.Node;
 import javafx.scene.Scene;
+import javafx.scene.control.CheckBox;
 import javafx.scene.control.ComboBox;
 import javafx.scene.control.Label;
 import javafx.scene.control.Separator;
+import javafx.scene.control.TextField;
 import javafx.scene.layout.HBox;
+import javafx.scene.layout.Priority;
 import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
 import org.controlsfx.control.CheckComboBox;
@@ -25,27 +28,30 @@ public class PhoneNumberFieldApp extends Application {
         if (c == null) {
             return null;
         }
-        CountryCallingCode code = (CountryCallingCode) c;
+        PhoneNumberField.CountryCallingCode code = (PhoneNumberField.CountryCallingCode) c;
         return "(+" + code.countryCode() + ") " + code.displayName("en");
     };
 
     @Override
     public void start(Stage stage) throws Exception {
-        PhoneNumberField view = new PhoneNumberField();
+        PhoneNumberField field = new PhoneNumberField();
 
         VBox controls = new VBox(10);
-        controls.getChildren().add(defaultCountrySelector(view));
-        controls.getChildren().add(preferredCountriesSelector(view));
+        addControl("Default Country", defaultCountrySelector(field), controls);
+        addControl("Available Countries", availableCountriesSelector(field), controls);
+        addControl("Preferred Countries", preferredCountriesSelector(field), controls);
+        addControl("Force Local Phone", useLocalPhoneNumberCheck(field), controls);
+        addControl("Mask", maskInputField(field), controls);
 
         VBox fields = new VBox(10);
-        addField(fields, "Number", view.phoneNumberProperty());
-        addField(fields, "Country", view.countryCodeProperty(), COUNTRY_CODE_CONVERTER);
-        addField(fields, "Local Number", view.localPhoneNumberProperty());
+        addField(fields, "Country", field.countryCodeProperty(), COUNTRY_CODE_CONVERTER);
+        addField(fields, "Number", field.phoneNumberProperty());
+        addField(fields, "Local Number", field.localPhoneNumberProperty());
 
         VBox vBox = new VBox(20);
         vBox.setPadding(new Insets(20));
         vBox.setAlignment(Pos.CENTER);
-        vBox.getChildren().addAll(controls, new Separator(), view, new Separator(), fields);
+        vBox.getChildren().addAll(controls, new Separator(), field, new Separator(), fields);
 
         Scene scene = new Scene(vBox, 500, 400);
 
@@ -57,23 +63,68 @@ public class PhoneNumberFieldApp extends Application {
     }
 
     private Node defaultCountrySelector(PhoneNumberField view) {
-        ComboBox<CountryCallingCode> comboBox = new ComboBox<>();
-        comboBox.getItems().addAll(CountryCallingCode.defaultValues());
+        ComboBox<PhoneNumberField.CountryCallingCode> comboBox = new ComboBox<>();
+        comboBox.getItems().addAll(PhoneNumberField.CountryCallingCode.defaultValues());
         comboBox.valueProperty().bindBidirectional(view.defaultCountryCodeProperty());
-        HBox hBox = new HBox(10);
-        hBox.setAlignment(Pos.CENTER_LEFT);
-        hBox.getChildren().addAll(new Label("Default Country: "), comboBox);
-        return hBox;
+        return comboBox;
+    }
+
+    private Node availableCountriesSelector(PhoneNumberField field) {
+        CheckBox allCountries = new CheckBox("All");
+
+        CheckComboBox<PhoneNumberField.CountryCallingCode> comboBox = new CheckComboBox<>();
+        comboBox.getItems().addAll(PhoneNumberField.CountryCallingCode.defaultValues());
+        comboBox.setPrefWidth(250);
+        comboBox.getCheckModel().getCheckedItems().addListener((InvalidationListener) observable -> field.getAvailableCountryCodes().setAll(comboBox.getCheckModel().getCheckedItems()));
+        comboBox.getCheckModel().checkAll();
+
+        allCountries.selectedProperty().addListener((obs, oldV, newV) -> {
+            if (newV) {
+                comboBox.getCheckModel().checkAll();
+                comboBox.setDisable(true);
+            } else {
+                comboBox.getCheckModel().clearChecks();
+                comboBox.setDisable(false);
+            }
+        });
+
+        allCountries.setSelected(true);
+
+        HBox box = new HBox(10);
+        box.getChildren().addAll(allCountries, comboBox);
+
+        return box;
     }
 
     private Node preferredCountriesSelector(PhoneNumberField view) {
-        CheckComboBox<CountryCallingCode> comboBox = new CheckComboBox<>();
-        comboBox.getItems().addAll(CountryCallingCode.defaultValues());
+        CheckComboBox<PhoneNumberField.CountryCallingCode> comboBox = new CheckComboBox<>();
+        comboBox.getItems().addAll(PhoneNumberField.CountryCallingCode.defaultValues());
+        comboBox.setPrefWidth(300);
         Bindings.bindContent(view.getPreferredCountryCodes(), comboBox.getCheckModel().getCheckedItems());
-        HBox hBox = new HBox(10);
+        return comboBox;
+    }
+
+    private Node maskInputField(PhoneNumberField field) {
+        TextField mask = new TextField();
+        mask.textProperty().bindBidirectional(field.maskProperty());
+        return mask;
+    }
+
+    private Node useLocalPhoneNumberCheck(PhoneNumberField field) {
+        CheckBox localCheck = new CheckBox();
+        localCheck.selectedProperty().bindBidirectional(field.forceLocalPhoneNumberProperty());
+        return localCheck;
+    }
+
+    private void addControl(String name, Node control, VBox controls) {
+        Label label = new Label(name);
+        label.setPrefWidth(150);
+        HBox hBox = new HBox();
         hBox.setAlignment(Pos.CENTER_LEFT);
-        hBox.getChildren().addAll(new Label("Preferred Countries: "), comboBox);
-        return hBox;
+        hBox.getChildren().addAll(label, control);
+        HBox.setHgrow(label, Priority.NEVER);
+        HBox.setHgrow(control, Priority.ALWAYS);
+        controls.getChildren().add(hBox);
     }
 
     private void addField(VBox fields, String label, ObservableValue property) {
@@ -88,9 +139,15 @@ public class PhoneNumberFieldApp extends Application {
             value.textProperty().bind(Bindings.createStringBinding(() -> converter.apply(property.getValue()), property));
         }
 
-        HBox hBox = new HBox(10);
+        Label myLabel = new Label(label + ": ");
+        myLabel.setPrefWidth(100);
+
+        HBox hBox = new HBox();
         hBox.setAlignment(Pos.CENTER_LEFT);
-        hBox.getChildren().addAll(new Label(label + ": "), value);
+        hBox.getChildren().addAll(myLabel, value);
+        HBox.setHgrow(myLabel, Priority.NEVER);
+        HBox.setHgrow(value, Priority.ALWAYS);
+
         fields.getChildren().add(hBox);
     }
